requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-SocketIO==5.1.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
Flask-WTF==0.15.1
geopy==2.2.0
Pillow==8.3.2
python-socketio==5.4.0
eventlet==0.33.0
googlemaps==4.6.0
